dialog working as intended!
... except for flipping the tail depending on side of room characters are in,
for now.

lower and upper limits are generated for the tails position on every draw call
of the dialog bubble (ie when a character is appended), and they are used to
clamp down the tail's position when DialogController calls set_player_offset.

Also DialogBubble is preloaded once and instanced on-demand in DialogController